requirements.txt
Click==7.0
Flask==1.0.2
Flask-Caching==1.7.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
honcho==1.0.1
itsdangerous==1.1.0
Jinja2==2.10.1
json-logging-py==0.2
MarkupSafe==1.1.1
python-envcfg==0.2.0
python-json-logger==0.1.11
python3-logstash==0.4.80
SQLAlchemy==1.3.3
Werkzeug==0.15.4
python-dotenv==0.11.0
python-json-logger==0.1.11
python3-logstash==0.4.80
redis==3.2.1
